+++ scripts/adjust_matrix_coordinates.py
@@ -2,23 +2,28 @@
 """ 
 Script that takes a matrix of counts
 where the columns are genes and the rows
-are the spot coordinates like: 
+are the spot coordinates in a format like this:
  
     gene    gene
 XxY
 XxY
 
 And then removes the spots that are not present in 
-a tab delimited coordinates file that has either 4
+the spot coordinates file or if the under_tissue flag is 0
+The format of the spot coordinates file can be like this:
 
 x y new_x new_y 
 
-or 6 columns
+or
 
 x y new_x new_y pixel_x pixel_y
 
+or
+
+x y new_x new_y pixel_x pixel_y under_tissue
+
 Optionally, the coordinates of the spots in the matrix
-can be changed to the new coordinates (pixel or array).
+can be changed to the adjusted new coordinates (array).
 
 @Author Jose Fernandez Navarro <jc.fernandez.navarro@gmail.com>
 """
@@ -28,7 +33,7 @@ import sys
 import os
 import pandas as pd
 
-def main(counts_matrix, coordinates_file, update_coordinates, outfile, outformat):
+def main(counts_matrix, coordinates_file, update_coordinates, outfile):
 
     if not os.path.isfile(counts_matrix) or not os.path.isfile(coordinates_file):
         sys.stderr.write("Error, input file not present or invalid format\n")
@@ -42,22 +47,15 @@ def main(counts_matrix, coordinates_file, update_coordinates, outfile, outformat
     with open(coordinates_file, "r") as filehandler:
         for line in filehandler.readlines():
             tokens = line.split()
-            assert(len(tokens) == 6 or len(tokens) == 4)
+            assert(len(tokens) == 6 or len(tokens) == 4 or len(tokens) == 7)
             if tokens[0] != "x":
                 old_x = int(tokens[0])
                 old_y = int(tokens[1])
-                new_x = round(float(tokens[2]),2)
-                new_y = round(float(tokens[3]),2)
-                if outformat == "array":
-                    new_coordinates[(old_x, old_y)] = (new_x,new_y)
-                elif len(tokens) == 6:
-                    pixel_x = float(tokens[4])
-                    pixel_y = float(tokens[5])
-                    new_coordinates[(old_x, old_y)] = (pixel_x, pixel_y)
-                else:
-                    sys.stderr.write("Error, output format is pixel coordinates but\n "
-                                     "the coordinates file only contains 4 columns\n")
-                    sys.exit(1)                    
+                new_x = round(float(tokens[2]), 2)
+                new_y = round(float(tokens[3]), 2)
+                if len(tokens) == 7 and not bool(tokens[6]):
+                    continue
+                new_coordinates[(old_x, old_y)] = (new_x, new_y)
     # Read the data frame (spots as rows)
     counts_table = pd.read_table(counts_matrix, sep="\t", header=0, index_col=0)
     new_index_values = list()
@@ -68,7 +66,7 @@ def main(counts_matrix, coordinates_file, update_coordinates, outfile, outformat
         x = int(tokens[0])
         y = int(tokens[1])
         try:
-            new_x, new_y = new_coordinates[(x,y)]
+            new_x, new_y = new_coordinates[(x, y)]
             if not update_coordinates:
                 new_x, new_y = x,y
             new_index_values.append("{0}x{1}".format(new_x,new_y))
@@ -87,19 +85,16 @@ def main(counts_matrix, coordinates_file, update_coordinates, outfile, outformat
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description=__doc__,
                                      formatter_class=argparse.RawDescriptionHelpFormatter)
-    parser.add_argument("--counts-matrix", required=True,
-                        help="Matrix with gene counts (genes as columns)")
+    parser.add_argument("counts_matrix",
+                        help="Matrix with gene counts (genes as columns) in TSV format")
     parser.add_argument("--outfile", help="Name of the output file")
     parser.add_argument("--update-coordinates", action="store_true", default=False,
                         help="Updates the spot coordinates in the output matrix with the\n"
                         "new coordinates present in the coordinates file")
     parser.add_argument("--coordinates-file",  required=True,
                         help="New coordinates in a tab delimited file")
-    parser.add_argument("--outformat", default="array",
-                        help="Output array positions or pixel positions"
-                        " array or pixel", choices=["array", "pixel"])
     args = parser.parse_args()
 
-    main(args.counts_matrix, args.coordinates_file, args.update_coordinates, args.outfile, args.outformat)
+    main(args.counts_matrix, args.coordinates_file, args.update_coordinates, args.outfile)