+++ scripts/convertEnsemblToNames.py
@@ -1,11 +1,9 @@
 #! /usr/bin/env python
 """
-Script that parses a ST data file generated
+Script that parses a Spatial Transcriptimics (ST) data file generated
 with the pipeline in matrix (TSV) format where the genes are named
-with ENSEMBL IDs and generates a new file
-with the ENSEMBL IDs converted to gene names IDS.
-For that the script also needs the annotation file (GFF format) used to create
-the ST dataset.
+with ENSEMBL IDs and generates a new file with the ENSEMBL IDs converted to gene names.
+The script needs the annotation file (GFF format) used to create the ST dataset.
 
 @Author Jose Fernandez Navarro <jc.fernandez.navarro@gmail.com>
 """
@@ -75,11 +73,12 @@ def main(st_data_file, annotation, output_file):
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description=__doc__,
                                      formatter_class=argparse.RawDescriptionHelpFormatter)
-    parser.add_argument("st_data_file", help="ST data file in TSV format")
+    parser.add_argument("counts_matrix", 
+                        help="Matrix with gene counts (genes as columns) in TSV format")
     parser.add_argument("--output", default="output.tsv",
                         help="Name of the output file, default output.tsv")
     parser.add_argument("--annotation", required=True,
                         help="Path to the annotation file used to generate the data")
     args = parser.parse_args()
-    main(args.st_data_file, args.annotation, args.output)
+    main(args.counts_matrix, args.annotation, args.output)